scx_bpfland: introduce dynamic nvcsw threshold
Instead of using a fixed value to classify tasks based on their average
amount of voluntary context switches, try to periodically evaluate an
optimal threshold based on a global average of all the tasks.

Tasks with an average amount of voluntary context switches greater than
the global average will be classified as interactive.

This enables a more precise adjustment of classification criteria to
swiftly respond to global system changes. Tasks are quickly classified
as interactive, but if the system experiences too many interactive
events, the criteria for maintaining interactive status become stricter.

This creates a natural selection process where only the most deserving
tasks remain interactive.

Additionally, introduce the new option `--nvcsw-max-thresh N`, which
allows to extend or restrict the fluctuation range of the global average
threshold for voluntary context switches.

Signed-off-by: Andrea Righi <righi.andrea@gmail.com>